Revert "Improve config formatting"
This reverts commit 3d0bac0df2df8565303e69aef7d4159a39d5cb7d.

Uncomment again commented out value and bump version.

Comment by Paul Wouters:
The value of 3072 was tailored to cause a failure for ANY requries to isc.org,
which are used a lot by attackers. Now with 4096,
it will fit and the query can be abused again to
cause amplification with that popular dns query.